+++ CHANGELOG.md
@@ -34,6 +34,10 @@
 
 * `build_cellranger_reference`: Bump go version to `1.21.4` when building seqkit for testing the component (PR #624).
 
+## BUG FIXES
+
+* `dataflow/concatenate_h5mu` and `dataflow/concat`: Fix an issue where joining columns with different datatypes caused `TypeError` (PR #619).
+
 # openpipelines 0.12.1
 
 ## BUG FIXES